requirements.txt
# Video Intelligence System - Dependencies
# PyTorch (install separately with CUDA support)
# pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu121

# AI/ML Libraries
transformers>=4.36.0
openai-whisper>=20231117
open-clip-torch>=2.23.0
insightface>=0.7.3
easyocr>=1.7.0
opencv-python>=4.8.1
scikit-learn>=1.3.2
onnxruntime-gpu>=1.16.0

# Database & Data
sqlalchemy>=2.0.23
pandas>=2.1.4

# Video Processing
ffmpeg-python>=0.2.0
Pillow>=10.1.0

# Utilities
tqdm>=4.66.1
pyyaml>=6.0.1
numpy>=1.24.0
